Set Ruby 2.1 as minimum version supported by Profiler
We've decided not to support Ruby 2.0 in the first release of the
profiler.

I've piggy-backed on the existing mechanism for only loading the
profiler if it is supported, and thus, on Ruby 2.0 is never even
loaded (allowing us to use 2.1 features in code).

This also enables us to simplify the CPU extension support: we no
longer have to separately check if we're on 2.1 or not, and we can
now assume that if the extension is being loaded, then we're on a
supported Ruby version.

I've tested that `ddtracerb exec` also continues to work correctly
on 2.0 (it just doesn't start profiling, even if you try to
enable it).