Avoid hangs when forked process outputs too much to stderr
Regular pipes have buffering limits, and when the buffer fills up,
the writer will just hang.

With our previous implementation, we were waiting for the forked
process to finish (with `Process.wait`) before we actually read
anything from the pipe. This worked if the output of the process
was enough to fit in the buffer, but when there was too much output
(in my case, the output of a Ruby segfault), then we'd get into a
deadlock: the forked process did not finish because it was trying
to write to the pipe, and the supervisor process did not read from
the pipe because it was waiting for the forked process to finish.

By switching to an actual temporary file, this issue no longer
happens, as the forked process can write as much as it wants, as
long as there's space in the temporary files folder.