Workaround Ruby segfault when reading instance variable from Process::Waiter
In Ruby versions 2.3 to 2.6, there's a very special subclass of
`Thread` called `Process::Waiter` that causes VM crashes whenever
something tries to read its instance variables. This subclass of thread
only shows up when the `Process.detach` API gets used.

In this case the "something that tries to read its instance variables"
is our stack sampler, which was trying to read the `native_thread_id`
from one such instance.

I've reported this issue upstream https://bugs.ruby-lang.org/issues/17807

To workaround this issue, I've added a check if the instance variable
exists before reading it, and validated it works with all affected
Ruby versions (as well as adding regression tests for it all).